Fix seek directions in notification. Fix duplicates. Returned to FMService.
